pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Feofilaktov, Sergey <man@smairon.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
greenlet = "^3.0.3"
fire = "^0.5.0"
envparse = "^0.2.0"
fastapi = "^0.103.1"
pydantic-settings = "^2.0.3"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.7"
faker = "^19.6.0"
uvicorn = "^0.23.2"
pydantic-extra-types = "^2.0.0"
pyjwt = "^2.8.0"
orjson = "^3.9.7"
httpx = "^0.25.1"
zodchy = "^0.1.1"
sqlalchemy-schema-factory = "^0.1.4"
pancho = "^1.0.0"
alembic-tools = "^0.1.0"
streamlord = "^0.1.0"
zorge = "^1.1.1"
hermitage = "^0.1.4"
hermitage-alchemy = "^0.1.3"
zodchy-security = "^0.1.0"
zodchy-patterns = "^0.2.0"
zodchy-fastapi = "^0.1.0"
zodchy-identity = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
